requirements.txt
numpy
scipy
scikit-learn
matplotlib
tensorflow-gpu
opencv-python
tqdm
jupyter